Fix w64 warnings

+++ modules/text/src/ocr_hmm_decoder.cpp
@@ -934,7 +934,7 @@ OCRHMMClassifierCNN::OCRHMMClassifierCNN (const string& filename)
 
     nr_feature  = weights.rows;
     nr_class    = weights.cols;
-    patch_size  = sqrt(kernels.cols);
+    patch_size  = (int)sqrt(kernels.cols);
     // algorithm internal parameters
     window_size = 32;
     num_quads   = 25;
